feat(image): Move to assets folder
