+++ packages/astro/src/assets/README.md
