+++ packages/astro/src/assets/index.ts
