+++ packages/astro/src/assets/internal.ts
