+++ packages/astro/src/assets/services/sharp.ts
