+++ packages/astro/src/assets/services/squoosh.ts
