+++ packages/astro/src/assets/types.ts
