+++ packages/astro/src/assets/utils/metadata.ts
