+++ packages/astro/src/assets/utils/queryParams.ts
