+++ packages/astro/src/assets/vite-plugin-assets.ts
